+++ src/index.js
@@ -1,10 +1,3 @@
-export {
-  createReducer,
-  handle,
-  getState,
-  map,
-  combineReducers,
-} from './createReducer'
+export { createReducer, getState, map, combine } from './createReducer'
 export { createAction } from './createAction'
-export { asId } from './model'
 export { createStore } from './createStore'