+++ src/index.ts
@@ -1,10 +1,8 @@
 export {
   createReducer,
-  handle,
   getState,
   map,
-  combineReducers,
+  combine,
 } from './createReducer'
 export { createAction } from './createAction'
-export { asId } from './model'
 export { createStore } from './createStore'