+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## 3.0.0
+
+New API for best static type inference
+
+```js
+createReducer(
+  'name',
+  initialState,
+  handle => [
+    handle(action, (state, payload) => ...)
+    handle(reducer1, (state, reducer1State) => ...)
+  ],
+)
+```
+
+## 2.0.0
+
+`redux-steroid` -> `flaxom`
+
 ## 1.7.0
 
 - Improve `store.getState()` API (getState without arguments now equal `getStateInternal`)